requirements.txt
flask==2.0.1
line-bot-sdk==3.16.2
python-dotenv==0.19.0 